Refactor: Improve project structure, configuration, and code quality
This commit addresses a broad request to analyze and improve the project by implementing a series of significant enhancements.

Key changes include:

- **Code Structure:** Refactored the monolithic `gui_main.py` by moving the `SettingsWindow` and `ReportBuilderWindow` classes into a new, more modular `gui/` package.
- **Configuration:** Centralized the configuration of required CSV column names by moving them from hardcoded values in `core.py` to a new `column_mappings` section in `config.json`.
- **Code Quality:**
    - Added comprehensive docstrings to key functions in `shopify_tool/core.py` and `shopify_tool/analysis.py` to improve readability and maintainability.
    - Refactored pandas code to eliminate 51/52 warnings, primarily by removing `inplace=True` and using safe assignment practices to avoid `SettingWithCopyWarning`.
- **Dependency Management:** Pinned all dependency versions in `requirements.txt` to ensure a stable and reproducible build environment.

These changes make the codebase more robust, flexible, and easier to maintain for future development. All existing tests pass with the new changes.
