+++ gui/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the 'gui' directory a Python package.
