+++ requirements.txt
@@ -1,10 +1,28 @@
-pandas
-numpy
-xlsxwriter
-xlrd
-xlutils
-pytest
-pytest-mock
-openpyxl
-pyinstaller
-customtkinter
+altgraph==0.17.4
+customtkinter==5.2.2
+darkdetect==0.8.0
+et_xmlfile==2.0.0
+greenlet==3.2.3
+iniconfig==2.1.0
+numpy==2.3.2
+openpyxl==3.1.5
+packaging==25.0
+pandas==2.3.1
+playwright==1.54.0
+pluggy==1.6.0
+pyee==13.0.0
+Pygments==2.19.2
+pyinstaller==6.15.0
+pyinstaller-hooks-contrib==2025.8
+pytest==8.4.1
+pytest-mock==3.14.1
+python-dateutil==2.9.0.post0
+pytz==2025.2
+setuptools==80.9.0
+six==1.17.0
+typing_extensions==4.14.1
+tzdata==2025.2
+xlrd==2.0.2
+xlsxwriter==3.2.5
+xlutils==2.0.0
+xlwt==1.3.0
